pyproject.toml
[project]
name = "igbo-archives-platform"
version = "1.0"
description = "Preserving Igbo history and culture"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django==6.0.2",
    "gunicorn==25.0.3",
    "whitenoise==6.11.0",
    "boto3==1.42.43",
    "django-storages==1.14.6",
    "pillow==12.1.0",
    "python-dotenv==1.2.1",
    "django-allauth==65.14.1",
    "django-threadedcomments==2.0",
    "django-pwa==2.0.1",
    "django-meta==2.5.0",
    "django-dbbackup==5.1.2",
    "django-taggit==6.1.0",
    "django-htmx==1.27.0",
    "django-editorjs-fields==0.2.7",
    "django-webpush==0.3.6",
    "huey==2.6.0",
    "bleach==6.3.0",
    "cryptography==46.0.4",
    "pyjwt==2.11.0",
    "google-genai==1.62.0",
    "groq==1.0.0",
    "gtts==2.5.4",
    "duckduckgo-search==8.1.1",
    "requests==2.32.5",
    "django-csp==4.0",
    "django-cleanup==9.0.0",
    "djangorestframework==3.16.0",
    "nh3>=0.3.2",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.2",
    "pytest-django==4.11.1",
    "ruff==0.15.0",
    "coverage==7.13.3",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]  # line length handled by formatter
